hypotesis_context/requirements.txt
Django~=3.0.6
djangorestframework~=3.11.0
PyMySQL~=0.9.3